@@ -14,21 +14,25 @@ classifiers=[
     'Intended Audience :: Developers',
     "License :: OSI Approved :: Apache Software License",
 ]
+packages = [
+    { include = "wagtail_spa_integration", from = "src" }
+]
 
 [tool.poetry.urls]
 "homepage" = "https://gitlab.com/thelabnyc/wagtail-spa-integration"
 
 [tool.poetry.dependencies]
-python = "^3.7"
-wagtail = "^4.1.0"
+python = "^3.10"
+wagtail = ">=5.2,<6.0"
 Django = ">=3.2.0"
-wagtail-headless-preview = "^0.2"
-django-filter = "^2.3.0"
+wagtail-headless-preview = "^0.8"
+django-filter = "^23.3"
 
 [tool.poetry.dev-dependencies]
 django-cors-headers = "^3.4.0"
-django-filter = "^2.3.0"
+flake8 = ">=3.3.0"
+tox = ">=2.7.0"
 
 [build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
+requires = [ "poetry-core>=1.0.0",]
+build-backend = "poetry.core.masonry.api"
