@@ -1,3 +1,5 @@
+# flake8: noqa
+
 from .base import *
 
 # SECURITY WARNING: don't run with debug turned on in production!
@@ -10,7 +12,6 @@ SECRET_KEY = "ig5hjl3_@rsp@^@0v1nzjs#$(*w@y2w_oc%)$6u$elyaef^+2c"
 ALLOWED_HOSTS = ["*"]
 
 EMAIL_BACKEND = "django.core.mail.backends.console.EmailBackend"
-
 
 try:
     from .local import *
